- Fix: compute mip levels after loading texture dimensions, not before.
- Refactor VAO, buffers and shape logic
- Fix: OnComponentAdded not receiving components by reference
- Make CMesh non copyable and CCollider copyable creating a new mesh
- Improve UUID class
- Refactor Asset Manager to use Handles and update engine logic so the only source of memory is AssetManager

+++ src/newleaf/components/meta/cUUID.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <map>
 #include <string>
 
@@ -9,7 +10,7 @@
 namespace nl {
 
 struct CUUID {
-    uint32_t uuid{};
+    uint64_t uuid{};
 
     void print() const { ENGINE_BACKTRACE("\t\tuuid: {}", uuid); }
